move property handling from Horn into its own class
